+++ src/WPBT/WPBT_Settings.php
@@ -36,6 +36,9 @@ class WPBT_Settings {
 	public function __construct( WP_Beta_Tester $wp_beta_tester, $options ) {
 		self::$options        = $options;
 		$this->wp_beta_tester = $wp_beta_tester;
+		if ( isset( self::$options['hide_report_a_bug'] ) ) {
+			add_filter( 'wpbt_hide_report_a_bug', '__return_true' );
+		}
 	}
 
 	/**
@@ -47,9 +50,11 @@ class WPBT_Settings {
 		$this->load_hooks();
 		( new WPBT_Core( $this->wp_beta_tester, self::$options ) )->load_hooks();
 		( new WPBT_Extras( $this->wp_beta_tester, self::$options ) )->load_hooks();
-		( new WPBT_Bug_Report( $this->wp_beta_tester, self::$options ) )->load_hooks();
 		( new WPBT_Extras( $this->wp_beta_tester, self::$options ) )->skip_autoupdate_email();
 		( new WPBT_Help() )->load_hooks();
+		if ( ! apply_filters( 'wpbt_hide_report_a_bug', false ) ) {
+			( new WPBT_Bug_Report( $this->wp_beta_tester, self::$options ) )->load_hooks();
+		}
 	}
 
 	/**
@@ -60,12 +65,15 @@ class WPBT_Settings {
 	public function load_hooks() {
 		add_action( 'admin_init', array( $this, 'add_settings' ) );
 		add_action( is_multisite() ? 'network_admin_menu' : 'admin_menu', array( $this, 'add_plugin_menu' ) );
-		add_action( 'admin_bar_menu', array( $this, 'admin_bar_menu' ), 80 );
 		add_action( 'network_admin_edit_wp_beta_tester', array( $this, 'update_settings' ) );
 		add_action( 'admin_init', array( $this, 'update_settings' ) );
 
 		add_action( 'admin_head-plugins.php', array( $this->wp_beta_tester, 'action_admin_head_plugins_php' ) );
 		add_action( 'admin_head-update-core.php', array( $this->wp_beta_tester, 'action_admin_head_plugins_php' ) );
+
+		if ( ! apply_filters( 'wpbt_hide_report_a_bug', false ) ) {
+			add_action( 'admin_bar_menu', array( $this, 'admin_bar_menu' ), 80 );
+		}
 	}
 
 	/**