+++ CHANGES.md
@@ -1,4 +1,6 @@
 [unreleased]
+
+#### 3.5.1 / 2023-07-06
 * add setting in `Extra Settings` to hide `Report a Bug`
 * add filter `wpbt_hide_report_a_bug`
 * update if `Report a Bug` plugin active